ABCL helped me uncover a tiny bug.
Sometimes the :name would be passed to the ensure-generic-function on the inner
gf, which would cause it to have the same name as the generic*, which would cause
it to set the fdefinition of that name to itself. But I guess other impls do not
change the fdefinition when add-method is called; they only reinstall the
funcallable-instance-function to the discriminating function, so I did not
notice. Additionally, it only happened in defgeneric*, not defmethod*, because
defmethod* never passes the :name argument in.